refactor: adjust code logic
